// src/domain/models/PacientesModel.ts
export interface PacienteModel {
    id:            number;
    breed:         string;
    names:         string;
    photo:         string;
    birthday:      Date;
    age:           string;
    sex:           string;
    color:         string;
    fur:           string;
    particularity: string;
    allergy:       string;
    status:        number;
    public_id:     string;
    client:        number;
}